fix(domain): apply the third Codex review of #15
- GitActivity::medianCorrectionCommitsAfterOpen and ::medianConcurrentBranches
  are now ?float: an even-sized PR sample can legitimately yield a fractional
  median (e.g. 1 and 2 -> 1.5). InterventionThresholds::levelForMedian() and
  ParallelismThresholds::levelForMedian() now take a float and read every band
  as "at least" its threshold, with "never"/"White" reserved for exactly 0 —
  anywhere strictly between 0 and 1 still counts as some correction or some
  concurrency. Boundary tests added for 1.5, 2.5 and 0.5.
- Notes are now a pointer-bearing Note(text, Pointer) value object, the same
  invariant Evidence enforces, instead of a bare string that could carry an
  unverifiable claim. AxisVerdict::$notes and Assessment::$notes both use it.
- Assessment carries the ?ProfileIdentity read for the profile (null only for
  a NotAssessable result whose gate broke before profile.json could be read),
  so the renderer has id/role for the output heading without reopening the
  profile source.
- Declarative keeps only presence (bool $present): the raw declaratif.md text
  is never carried past the gate, since spec 05 reserves quoting that
  questionnaire for the README and docs/methode.md, never for rendered
  assessment output. Unrelated to ProfileIdentity::$note (profile.json).

+++ src/Domain/Profile/GitActivity.php
@@ -17,11 +17,11 @@ final readonly class GitActivity
         public ?int $pullRequestsTotal,
         public ?float $medianFilesChanged,
         public ?float $medianLinesChanged,
-        public ?int $medianCorrectionCommitsAfterOpen,
+        public ?float $medianCorrectionCommitsAfterOpen,
         public ?int $mergedWithoutHumanEditAfterOpen,
         public ?float $aiCoauthoredRatio,
         public ?int $maxConcurrentBranches,
-        public ?int $medianConcurrentBranches,
+        public ?float $medianConcurrentBranches,
         public ?ContextFiles $contextFiles,
     ) {
     }
